+++ swift/Sources/LibSignalClient/AsyncUtils.swift
@@ -64,6 +64,10 @@ extension SignalCPromiseFfiCheckSvr2CredentialsResponse: PromiseStruct {
     typealias Result = SignalFfiCheckSvr2CredentialsResponse
 }
 
+extension SignalCPromiseMutPointerRegisterAccountResponse: PromiseStruct {
+    typealias Result = SignalMutPointerRegisterAccountResponse
+}
+
 /// A type-erased version of ``Completer``.
 ///
 /// Not for direct use, see Completer instead.