+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.72.1
 
+- Bridge registration service client to Swift.